+++ feature/cat-editor/src/main/java/com/dede/android_eggs/cat_editor/ColorPaletteDialog.kt
@@ -71,6 +71,7 @@ fun ColorPaletteDialog(
     val performColorSelected by rememberUpdatedState(onColorSelected)
 
     ModalBottomSheet(
+        modifier = Modifier.sizeIn(maxWidth = 420.dp),
         sheetState = sheetState,
         onDismissRequest = {
             visible = false